refactor: extract Nutzap profile event helpers
